Fix syntax, align with roadmap, refactor auth and UI
