+++ package.json
@@ -27,7 +27,8 @@
     "jsonwebtoken": "^9.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^7.11.0",
+    "@react-oauth/google": "^0.12.1",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "^5.0.1",
     "styled-components": "^6.1.1"
   },
@@ -58,7 +59,6 @@
   },
   "overrides": {
     "nth-check": "^2.0.1",
-    "postcss": "^8.4.31",
-    "webpack-dev-server": "^5.2.0"
+    "postcss": "^8.4.31"
   }
 }
