+++ src/nlp/cd-detector.test.js
@@ -18,14 +18,25 @@ const mockTensor = {
 
 jest.mock('@tensorflow/tfjs', () => {
   const originalTf = jest.requireActual('@tensorflow/tfjs');
+  // Helper for mock factory (duplicated because factory is hoisted)
+  const createMockSqueezedTensorInside = (similarityValue = 0.9) => ({
+    neg: jest.fn(() => createMockSqueezedTensorInside(similarityValue)),
+    add: jest.fn(() => createMockSqueezedTensorInside(similarityValue)),
+    dataSync: jest.fn(() => [similarityValue]),
+    dispose: jest.fn(),
+  });
+
+  const mockTensorInside = {
+      squeeze: jest.fn(() => createMockSqueezedTensorInside()),
+      dispose: jest.fn(),
+  };
+
   return {
     ...originalTf,
-    mean: jest.fn((tensor) => mockTensor), // Always return mockTensor
-    metrics: {
-      ...originalTf.metrics,
-      cosineDistance: jest.fn((vec1, vec2) => createMockSqueezedTensor(0.9)), // Default high similarity
-    },
-    tensor2d: jest.fn(() => mockTensor),
+    mean: jest.fn((tensor) => mockTensorInside),
+    mul: jest.fn((t1, t2) => mockTensorInside),
+    sum: jest.fn((t) => createMockSqueezedTensorInside(0.9)),
+    tensor2d: jest.fn(() => mockTensorInside),
     equal: jest.fn(() => ({
       all: jest.fn(() => ({
         dataSync: jest.fn(() => [1]),
@@ -54,8 +65,8 @@ describe('detectCDs', () => {
   });
 
   test('should return an empty array if no cognitive distortions are detected', async () => {
-    // Simulate low similarity for all CDs
-    tf.metrics.cosineDistance.mockReturnValue(createMockSqueezedTensor(0.1));
+    // Simulate low similarity (0.1) which is below all thresholds
+    tf.sum.mockReturnValue(createMockSqueezedTensor(0.1));
     const text = 'This is a neutral sentence.';
     const result = await detectCDs(text);
     expect(result).toEqual([]);
@@ -63,21 +74,38 @@ describe('detectCDs', () => {
     expect(mockEmbed).toHaveBeenCalled();
   });
 
-  test('should detect "All-or-Nothing" distortion', async () => {
-    // Mock cosineDistance to return high similarity for the relevant CD type
-    // This requires knowing the order of CD types being checked in detectCDs
-    // For simplicity, we'll make all cosine distances high enough to trigger detection
-    // and then filter based on the expected outcome.
-    tf.metrics.cosineDistance.mockImplementation((vec1, vec2) => {
-      // In a real scenario, you might inspect vec1/vec2 to determine which CD is being checked
-      // For now, we'll just return a high similarity for all.
-      return createMockSqueezedTensor(0.95);
-    });
+  test('should detect "All-or-Nothing" distortion with high similarity', async () => {
+    // 0.95 > 0.6 (high sens), 0.7 (med sens), 0.8 (low sens)
+    tf.sum.mockReturnValue(createMockSqueezedTensor(0.95));
 
     const text = 'I always fail at everything.';
-    const result = await detectCDs(text, 'high');
+    const result = await detectCDs(text, 'high'); // high sensitivity = 0.6 threshold
     expect(result).toContain('All-or-Nothing');
     expect(use.load).toHaveBeenCalledTimes(1);
     expect(mockEmbed).toHaveBeenCalled();
+  });
+
+  test('should respect sensitivity thresholds', async () => {
+      // Thresholds: Low=0.8, Medium=0.7, High=0.6
+
+      // Case 1: Similarity 0.65.
+      // Should be detected at High (0.6) but NOT Medium (0.7) or Low (0.8)
+      tf.sum.mockReturnValue(createMockSqueezedTensor(0.65));
+
+      let result = await detectCDs('text', 'high');
+      expect(result.length).toBeGreaterThan(0); // 0.65 > 0.6
+
+      result = await detectCDs('text', 'medium');
+      expect(result.length).toBe(0); // 0.65 < 0.7
+
+      // Case 2: Similarity 0.75
+      // Should be detected at Medium (0.7) but NOT Low (0.8)
+      tf.sum.mockReturnValue(createMockSqueezedTensor(0.75));
+
+      result = await detectCDs('text', 'medium');
+      expect(result.length).toBeGreaterThan(0); // 0.75 > 0.7
+
+      result = await detectCDs('text', 'low');
+      expect(result.length).toBe(0); // 0.75 < 0.8
   });
 });
