+++ src/services/api.js
@@ -9,19 +9,17 @@ const apiClient = axios.create({
   },
 });
 
-// Optional: Add an interceptor to include JWT token if available
-// For now, it's not implemented as token management will be in AuthContext
-// apiClient.interceptors.request.use(
-//   (config) => {
-//     const token = localStorage.getItem('token'); // Or get from context
-//     if (token) {
-//       config.headers['x-access-token'] = token;
-//     }
-//     return config;
-//   },
-//   (error) => {
-//     return Promise.reject(error);
-//   }
-// );
+apiClient.interceptors.request.use(
+  (config) => {
+    const token = localStorage.getItem('token');
+    if (token) {
+      config.headers['x-access-token'] = token;
+    }
+    return config;
+  },
+  (error) => {
+    return Promise.reject(error);
+  }
+);
 
 export default apiClient;
